backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.10.0
pydantic-settings>=2.1.0
yt-dlp>=2025.1.16
google-generativeai>=0.3.0
openai-whisper>=20231117
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
aiofiles>=23.2.1
